requirements.txt
openai
flask
botbuilder-core
botframework-connector
